Add dsl context class for eval DSL in clean object
