+++ lib/dashdog.rb
@@ -6,6 +6,7 @@ require "dashdog/client"
 require "dashdog/converter"
 require "dashdog/actions"
 require "dashdog/cli"
+require "dashdog/dsl_context"
 
 module Dashdog
   # Your code goes here...
